refactor(utils): add getDiffInCalendarWeeks as ESM
